changed todo methods to Routers folder
